[xml] Change CentOffset to a floating point number

+++ src/Rocksmith2014.Conversion/ConvertInstrumental.fs
@@ -70,7 +70,7 @@ let sngToXml (attr: Manifest.Attributes option) (sng: SNG) =
         match attr with
         | Some attr ->
             let m = MetaData(Arrangement = attr.ArrangementName,
-                             CentOffset = int (attr.CentOffset.GetValueOrDefault()),
+                             CentOffset = float32 (attr.CentOffset.GetValueOrDefault()),
                              AverageTempo = attr.SongAverageTempo.GetValueOrDefault(),
                              Title = attr.SongName,
                              TitleSort = attr.SongNameSort,
